requirements.txt
dnspython==2.1.0
discord.py==1.6.0
PyNaCl==1.4.0
youtube_dl==2021.3.3
youtube_search==2.1.0
youtube_title_parse==1.0.0
lyricsgenius==3.0.0
PyMongo==3.11.3
emoji==1.2.0